rrs/backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
django-cors-headers
channels
channels-redis
python-dotenv
psycopg2-binary